AT-6380 reorganizing the external methods and modals, and fixing a small style bug
