+++ src/sass/components/_permissions_modal.scss
@@ -39,7 +39,7 @@
     order: 1;
     align-self: stretch;
     flex-grow: 0;
-    margin: 8px 0px;
+    margin: 0px;
     display: flex;
     flex-direction: column;
     align-items: flex-start;
