+++ src/store/validation.modules.ts
@@ -1,0 +1,5 @@
+export const validEmail = (email: string): boolean => {
+  const isValidEmail =
+    /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/g;
+  return isValidEmail.test(email);
+};
