+++ types/Actions.ts
@@ -1,0 +1,8 @@
+export interface KeyValuePair {
+    [key: string]: any;
+}
+  
+export interface ActionObject {
+    action: string;
+    data?: KeyValuePair;
+}
